[heft] Remove LifecycleOperationRunner
