+++ apps/heft/src/pluginFramework/HeftLifecycle.ts
@@ -22,7 +22,6 @@ import {
   type IHeftLifecycleSession
 } from './HeftLifecycleSession';
 import type { ScopedLogger } from './logging/ScopedLogger';
-import type { LifecycleOperationRunnerType } from '../operations/runners/LifecycleOperationRunner';
 
 export interface IHeftLifecycleContext {
   lifecycleSession?: HeftLifecycleSession;
@@ -39,7 +38,7 @@ export class HeftLifecycle extends HeftPluginHost {
     HeftLifecyclePluginDefinition,
     IHeftLifecyclePlugin<object | void>
   > = new Map();
-  private readonly _lifecycleLoggersByType: Map<string | undefined, ScopedLogger> = new Map();
+  private _lifecycleLogger: ScopedLogger | undefined;
 
   private _isInitialized: boolean = false;
 
@@ -181,13 +180,11 @@ export class HeftLifecycle extends HeftPluginHost {
     }
   }
 
-  public getLifecycleLoggerByType(type: LifecycleOperationRunnerType | undefined): ScopedLogger {
-    let logger: ScopedLogger | undefined = this._lifecycleLoggersByType.get(type);
+  public get lifecycleLogger(): ScopedLogger {
+    let logger: ScopedLogger | undefined = this._lifecycleLogger;
     if (!logger) {
-      logger = this._internalHeftSession.loggingManager.requestScopedLogger(
-        `lifecycle${type === undefined ? '' : `:${type}`}`
-      );
-      this._lifecycleLoggersByType.set(type, logger);
+      logger = this._internalHeftSession.loggingManager.requestScopedLogger(`lifecycle`);
+      this._lifecycleLogger = logger;
     }
     return logger;
   }
